+++ client/components/reviewMod/star_rating.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const StarRating = (rating, i) => (
+  <div key={i}>
+    {console.log(rating)}
+  </div>
+);
+
+export default StarRating;
